requirements.txt
aniso8601==9.0.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
Flask==2.1.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
protobuf==4.21.1
pycparser==2.21
PyJWT==2.4.0
PyMySQL==1.0.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.37
urllib3==1.26.9
Werkzeug==2.1.2
